+++ utils_test.go
@@ -5,10 +5,10 @@ import (
 )
 
 func TestUtilsGetFilesSortedWeights(t *testing.T) {
-	files := make(map[int32]*File)
+	files := make(map[int64]*File)
 	var expectedTotalWeight float64
 	for i := 5; i > 0; i-- {
-		files[int32(i)] = NewFile(NewFileReadCloser(nil, 0, nil), int64(i))
+		files[int64(i)] = NewFile(NewFileReadCloser(nil, 0, nil), int64(i))
 		expectedTotalWeight += 1.0 / float64(i)
 	}
 
@@ -26,10 +26,10 @@ func TestUtilsGetFilesSortedWeights(t *testing.T) {
 }
 
 func TestUtilsGetFilesSortedWeightsEmptyOnFinish(t *testing.T) {
-	files := make(map[int32]*File)
+	files := make(map[int64]*File)
 	for i := 5; i > 0; i-- {
 		file := NewFile(NewFileReadCloser(nil, 0, nil), 0)
-		files[int32(i)] = file
+		files[int64(i)] = file
 	}
 
 	weights, totalWeights := getFilesSortedWeights(files)
